Flask_server/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.1
Flask-RESTful==0.3.10
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
pathlib==1.0.1
pluggy==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.4
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
Werkzeug==3.0.1
wheel==0.41.2
sshtunnel>=0.4.0
paramiko>=2.11.0
